# software/admin/docs/swagger.yaml
definitions:
  dto.ApiRequest:
    properties:
      apiGroup:
        description: api组
        type: string
      createdBy:
        type: integer
      description:
        description: api中文描述
        type: string
      handle:
        type: string
      method:
        type: string
      path:
        description: api路径
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - apiGroup
    - createdBy
    - description
    - handle
    - method
    - path
    - title
    - type
    type: object
  dto.ApiSearchParams:
    properties:
      apiGroup:
        type: string
      desc:
        type: boolean
      method:
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      path:
        type: string
      title:
        type: string
    type: object
  dto.DeptRequest:
    properties:
      createdBy:
        type: integer
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 部门排序
        type: integer
      status:
        description: 状态
        type: integer
    required:
    - createdBy
    - leader
    - name
    - sort
    - status
    type: object
  dto.DeptSearchParams:
    properties:
      name:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      status:
        type: integer
    type: object
  dto.IDUriRequest:
    properties:
      id:
        type: integer
    type: object
  dto.PostRequest:
    properties:
      code:
        description: 岗位代码
        type: string
      createdBy:
        type: integer
      name:
        description: 岗位名称
        type: string
      remark:
        description: 描述
        type: string
      sort:
        description: 岗位排序
        type: integer
      status:
        description: 状态
        type: integer
    required:
    - code
    - createdBy
    - name
    - sort
    - status
    type: object
  dto.PostSearchParams:
    properties:
      code:
        type: string
      name:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      status:
        type: integer
    type: object
  po.Api:
    properties:
      apiGroup:
        description: api组
        type: string
      createBy:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      description:
        description: api中文描述
        type: string
      handle:
        type: string
      id:
        description: 主键ID
        type: integer
      method:
        description: 方法:创建POST|查看GET(默认)|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      title:
        type: string
      type:
        type: string
      updateBy:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  po.Post:
    properties:
      code:
        description: 岗位代码
        type: string
      createBy:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 岗位名称
        type: string
      remark:
        description: 描述
        type: string
      sort:
        description: 岗位排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  response.Data:
    properties:
      code:
        description: 自定义返回码  0:表示正常
        type: integer
      data:
        description: 返回的具体数据
      message:
        description: 关于这次响应的说明信息
        type: string
      meta:
        description: 数据meta
      namespace:
        description: 异常的范围
        type: string
      reason:
        description: 异常原因
        type: string
      recommend:
        description: 推荐链接
        type: string
      request_id:
        description: 请求Id
        type: string
      type:
        description: 数据类型, 可以缺省
        type: string
    type: object
  vo.ApiList:
    properties:
      count:
        type: integer
      result:
        items:
          $ref: '#/definitions/po.Api'
        type: array
    type: object
  vo.PostList:
    properties:
      count:
        type: integer
      result:
        items:
          $ref: '#/definitions/po.Post'
        type: array
    type: object
info:
  contact: {}
paths:
  /api/v1//blog/banner:
    get:
      consumes:
      - application/json
      description: 广告列表接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      summary: 广告列表
      tags:
      - Blog
  /api/v1/blog/advert:
    get:
      consumes:
      - application/json
      description: 广告列表接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      summary: 广告列表
      tags:
      - Blog
  /api/v1/system/api:
    get:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ApiSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, "data": [...], "message"=""}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/vo.ApiList'
              type: object
      security:
      - Bearer: []
      summary: api列表
      tags:
      - SysApi
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ApiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 创建基础api
      tags:
      - SysApi
  /api/v1/system/api/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 删除api
      tags:
      - SysApi
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: "api\t详情"
      tags:
      - SysApi
    put:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ApiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '}'''
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 更新api
      tags:
      - SysApi
  /api/v1/system/dept:
    delete:
      description: 删除数据
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 删除部门
      tags:
      - SysDept
    get:
      consumes:
      - application/json
      description: 部门列表
      parameters:
      - description: 名称, 状态
        in: body
        name: ddata
        required: true
        schema:
          $ref: '#/definitions/dto.DeptSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 部门列表数据
      tags:
      - SysDept
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      summary: 添加部门
      tags:
      - SysDept
  /api/v1/system/dept/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 部门详情
      tags:
      - SysDept
    put:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DeptRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 修改部门
      tags:
      - SysDept
  /api/v1/system/dept/tree:
    get:
      consumes:
      - application/json
      description: 部门所有数据树
      parameters:
      - description: 名称, 状态
        in: body
        name: ddata
        required: true
        schema:
          $ref: '#/definitions/dto.DeptSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: 部门所有数据树
      tags:
      - SysDept
  /api/v1/system/post:
    get:
      consumes:
      - application/json
      parameters:
      - description: 岗位名称, 岗位代码, 岗位状态
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.PostSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, "data": [...], "message"=""}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/vo.PostList'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 岗位列表
      tags:
      - SysPost
    post:
      consumes:
      - application/json
      parameters:
      - description: 岗位名称, 岗位代码,岗位排序, 岗位状态, 岗位描述
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ApiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - ApiKeyAuth: []
      summary: 创建岗位
      tags:
      - SysPost
  /api/v1/system/post/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - ApiKeyAuth: []
      summary: 删除岗位
      tags:
      - SysPost
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.IDUriRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - ApiKeyAuth: []
      summary: 岗位详情
      tags:
      - SysPost
    put:
      consumes:
      - application/json
      parameters:
      - description: 岗位名称, 岗位代码,岗位排序, 岗位状态, 岗位描述
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.PostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '}'''
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - ApiKeyAuth: []
      summary: 更新岗位
      tags:
      - SysPost
swagger: "2.0"